requirements/base.txt
#####################
# Base requirements #
#####################

Django==1.7.4
Unipath==1.1
argparse==1.3.0
django-appconf==1.0.1
django-extensions==1.5.0
django-filter==0.11.0
djangorestframework-jwt==1.7.2
django-model-utils==2.3.1
djangorestframework==2.4.4
path.py==8.1
python-dateutil==2.4.0
pytz==2014.10
simplejson==3.6.5
six==1.9.0

# Docs
docutils==0.12
Markdown==2.5.2
Sphinx==1.2.3
